Move maps api to own file
